Per review, moved the findRootDomain method to utils out of the userId modules
